Fixed almost all warnings from GCC
Various conversions to unsigned char *, sockaddr union in dht code
fixes strict aliasing issues.  The only thing left is unused
variables in the dht callback, which I think we can all agree
is not a big deal.  Also, scons changes to compile dht lib and
link it with the executable.